requirements.txt
PyYAML
imap-tools